backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
python-decouple==3.8
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
